src/requirements.txt
Flask
pymongo
python-dotenv
python-bsonjs